Export - fix

+++ manager/includes/extenders/export.class.inc.php
@@ -90,10 +90,14 @@ class EXPORT_SITE
 		elseif(!is_readable($directory)) return false;
 		else
 		{
-			foreach(glob($directory . '/*') as $path)
+			$files = glob($directory . '/*');
+			if(!empty($files))
 			{
-				if(is_dir($path)) $this->removeDirectoryAll($path);
-				else              $rs = unlink($path);
+    			foreach($files as $path)
+    			{
+    				if(is_dir($path)) $this->removeDirectoryAll($path);
+    				else              $rs = unlink($path);
+    			}
 			}
 		}
 		if($directory !== $this->targetDir) $rs = rmdir($directory);
